Move read_location, derives_from, check_graph
New core.py with read_location, derives_from, check_graph moved verbatim; main.py imports them back. No behavior change.
