+++ src/spdx3_validate/main.py
@@ -1,144 +1,18 @@
-# Copyright (c) 2024 Joshua Watt
-#
+# SPDX-FileContributor: Joshua Watt
+# SPDX-FileCopyrightText: 2024 Joshua Watt
+# SPDX-FileType: SOURCE
 # SPDX-License-Identifier: MIT
 
 import argparse
 import halo
 import json
 import jsonschema
-import pyshacl
 import rdflib
-import sys
-import textwrap
 import urllib.request
 
-from rdflib import RDF, RDFS, SH, URIRef
-
-from pathlib import Path
 from .version import VERSION
 from .spdx_versions import find_version, SPDX_VERSIONS
-
-
-def read_location(location):
-    if "://" in location:
-        with urllib.request.urlopen(location) as f:
-            return f.read()
-    elif location == "-":
-        return sys.stdin.read()
-    else:
-        with Path(location).open("r") as f:
-            return f.read()
-
-
-def derives_from(cls, target, shacl_graph):
-    if cls == target:
-        return True
-
-    for subclass in shacl_graph.objects(cls, RDFS.subClassOf):
-        if derives_from(subclass, target, shacl_graph):
-            return True
-
-    return False
-
-
-def check_graph(graph, shacl_graph, current_version, error_external):
-    errors = []
-
-    conforms, results, _ = pyshacl.validate(
-        graph,
-        shacl_graph=shacl_graph,
-        ont_graph=shacl_graph,
-    )
-
-    if not conforms:
-        results.bind("sh", SH)
-        nm = rdflib.namespace.NamespaceManager(results)
-
-        def norm(uri):
-            return nm.normalizeUri(uri)
-
-        def pnode(n):
-            if n:
-                return n.n3()
-            return "-"
-
-        # Collect all external map references
-        external_spdxids = set()
-        for spdxid in current_version.get_imports(graph):
-            # If the SpdxID is in the graph as a subject, than do
-            # not mark it as an external SpdxID, since there is a
-            # resolved definition for it
-            if (spdxid, None, None) in graph:
-                if error_external:
-                    errors.append(
-                        f"ERROR: {str(spdxid)} in an ExternalMap and also defined in the document"
-                    )
-            else:
-                external_spdxids.add(str(spdxid))
-
-        def check_external_ref_error(r):
-            if (r, RDF.type, SH.ValidationResult) not in results:
-                return False
-
-            if (r, SH.resultSeverity, SH.Violation) not in results:
-                return False
-
-            if (
-                r,
-                SH.sourceConstraintComponent,
-                SH.ClassConstraintComponent,
-            ) not in results:
-                return False
-
-            is_element = False
-            for ss in results.objects(r, SH.sourceShape):
-                if is_element:
-                    break
-
-                for cls in results.objects(ss, SH["class"]):
-                    is_element = derives_from(
-                        cls,
-                        URIRef(current_version.rdf_base + "Core/Element"),
-                        shacl_graph,
-                    )
-                    if is_element:
-                        break
-
-            if not is_element:
-                return False
-
-            for v in results.objects(r, SH.value):
-                if str(v) in external_spdxids:
-                    return True
-
-            return False
-
-        for report in results.subjects(RDF.type, SH.ValidationReport):
-            for r in results.objects(report, SH.result):
-                if check_external_ref_error(r):
-                    continue
-
-                e = []
-                e.append(
-                    f"Violation of type {norm(results.value(r, SH.sourceConstraintComponent))}:"
-                )
-                e.append(f"\tSeverity: {norm(results.value(r, SH.resultSeverity))}")
-                pg = rdflib.Graph()
-                pg += results.triples((results.value(r, SH.sourceShape), None, None))
-                if pg:
-                    e.append("\tSource Shape:")
-                    e.append(
-                        textwrap.indent(pg.serialize(format="ttl").strip(), "\t\t")
-                    )
-                e.append(f"\tFocus Node: {pnode(results.value(r, SH.focusNode))}")
-                e.append(f"\tValue Node: {pnode(results.value(r, SH.value))}")
-                e.append(f"\tResult path: {pnode(results.value(r, SH.resultPath))}")
-                e.append(f"\tMessage: {results.value(r, SH.resultMessage) or '-'}")
-                e.append("")
-
-                errors.append("\n".join(e))
-
-    return errors
+from .core import check_graph, read_location
 
 
 def iter_validation_errors(err):
